# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The local daemon's API (ADR 0010): what every surface that is not an agent reads and changes. It serves derived figures only, never prompt text, file contents or transcripts. Every route needs the bearer token from ~/.tokenops/daemon.url.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "TokenOps daemon API",
    "version": "local"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/routing/rules": {
      "post": {
        "operationId": "postRoutingRules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "delete": {
                    "type": "boolean"
                  },
                  "fallbacks": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "from_model": {
                    "type": "string"
                  },
                  "provider": {
                    "type": "string"
                  },
                  "quality": {
                    "type": "number"
                  },
                  "to_model": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "config": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    },
                    "routing_rules": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "Fallbacks": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "FromModel": {
                            "type": "string"
                          },
                          "Provider": {
                            "type": "string"
                          },
                          "Quality": {
                            "type": "number"
                          },
                          "ToModel": {
                            "type": "string"
                          },
                          "WhenClass": {
                            "type": "string"
                          },
                          "WhenWindowPctAbove": {
                            "type": "number"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Create, update or delete a routing rule"
      }
    },
    "/api/rules/analyze": {
      "get": {
        "operationId": "getRulesAnalyze",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Rule files analysed"
      }
    },
    "/api/rules/compress": {
      "get": {
        "operationId": "getRulesCompress",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Rule files compressed"
      }
    },
    "/api/rules/conflicts": {
      "get": {
        "operationId": "getRulesConflicts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Conflicts between rules"
      }
    },
    "/api/rules/inject": {
      "get": {
        "operationId": "getRulesInject",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "The rules to inject for a task"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:7878"
    }
  ]
}